upgrade @shopify/eslint-plugin 46.0.0
